Push

+++ Quarry.lua
@@ -140,7 +140,7 @@ function depositItems()
 	turtle.place()
 	for i = 1, NUM_SLOTS do
 		for _, v in pairs(fuelItems) do
-			if turtle.getItemDetail(i)['name'] ~= v then
+			if turtle.getItemDetail(i)['name'] ~= v and turtle.getItemDetail(i)['name'] ~= nil then
 				turtle.select(i)
 				turtle.drop()
 			end
